+++ HomeService/HomeService.service/Implementations/TeamService.cs
@@ -4,6 +4,7 @@ using HomeService.core.Models;
 using HomeService.service.Dtos;
 using HomeService.service.Dtos.TeamDto;
 using HomeService.service.Dtos.TeamRepository;
+using HomeService.service.Exeptions;
 using HomeService.service.Extentions;
 using HomeService.service.Interfaces;
 using Microsoft.AspNetCore.Hosting;
@@ -30,7 +31,10 @@ namespace HomeService.service.Implementations
         public async Task CreateAsync(TeamPostDto postDto)
         {
             Team team = _mapper.Map<Team>(postDto);
+
+            if(postDto.FormFile!=null)
             team.Image = postDto.FormFile.SaveImage(_env.WebRootPath, "assets/images/team");
+
             await _unitOfWork.TeamRepository.AddAsync(team);
             await _unitOfWork.CommitAsync();
 
@@ -39,6 +43,10 @@ namespace HomeService.service.Implementations
         public async Task<TeamGetDto> Get(int id)
         {
             Team team = await _unitOfWork.TeamRepository.GetAsync(x => x.Id == id, "Position");
+
+            if (team == null)
+                throw new ItemNotFoundExeption("Item is not found");
+
             TeamGetDto getDto = _mapper.Map<TeamGetDto>(team);
             return getDto;
         }
@@ -56,14 +64,21 @@ namespace HomeService.service.Implementations
         public async Task RemoveAync(int id)
         {
             Team team = await _unitOfWork.TeamRepository.GetAsync(x => x.Id == id, "Position");
-             _unitOfWork.TeamRepository.Remove(team);
+            if (team == null)
+                throw new ItemNotFoundExeption("Item is not found");
+
+            _unitOfWork.TeamRepository.Remove(team);
             await _unitOfWork.CommitAsync();
         }
 
         public async Task Update(int id, TeamPostDto postDto)
         {
             Team team = await _unitOfWork.TeamRepository.GetAsync(x => x.Id == id, "Position");
-           team.InstagramLink=postDto.InstagramLink;
+
+            if (team == null)
+                throw new ItemNotFoundExeption("Item is not found");
+
+            team.InstagramLink=postDto.InstagramLink;
            team.FacebookLink=postDto.FacebookLink;
            team.TwitterLink=postDto.TwitterLink;
            team.FullName=postDto.FullName;
